Convert usePullTreePaths to ts
